Handling of additional cases for COC and COF annotations, turn on by default in somatic caller

+++ src/com/rtg/vcf/annotation/ContraryObservationCountAnnotation.java
@@ -46,7 +46,7 @@ public class ContraryObservationCountAnnotation extends ContraryObservationFract
   }
 
   @Override
-  protected Object getValue(final long ref, final long alt) {
+  protected Object getValue(final int ref, final int alt) {
     return alt;
   }
